apps/api/requirements.txt
flask==3.0.3
prometheus-client==0.20.0
pika==1.3.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
